commit ter

+++ src/main/java/com/senai/rental_project/controllers/UsuarioController.java
@@ -8,13 +8,12 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.senai.rental_project.moldels.Usuario;
 import com.senai.rental_project.services.UsuarioService;
-
-import io.swagger.v3.oas.annotations.parameters.RequestBody;
 
 @RestController
 @RequestMapping("/usuario")
@@ -34,7 +33,7 @@ public class UsuarioController {
     }
     
     @DeleteMapping("/deletar-usuario/{id}")
-    public String deletarUsuario(@PathVariable Integer id) {
+    public String deletarUsuario(@PathVariable Long id) {
         if(usuarioService.deletarUsuario(id)) {
             return "Usuário removido com sucesso";
         }
@@ -42,7 +41,7 @@ public class UsuarioController {
     }
 
     @GetMapping("/buscar-usuarios/{id}")
-    public Usuario buscUsuario(@PathVariable Integer id) {
+    public Usuario buscUsuario(@PathVariable Long id) {
         return usuarioService.buscarUsuario(id);
     }
 
@@ -52,7 +51,7 @@ public class UsuarioController {
     }
 
     @PutMapping("/atualizar-usuario/{id}")
-    public String atualizarUsuario(@PathVariable Integer id, @RequestBody Usuario usuario) {
+    public String atualizarUsuario(@PathVariable Long id, @RequestBody Usuario usuario) {
         if(usuarioService.atualizarUsuario(id, usuario) != null) {
             return "Usuário atualizado com sucesso.";
         }
